Most Countries, new buttons
I need help with  showing multiple tables with Mapbox, I cant find anything on internet. I made, but not pushed, 2 working buttons for Pop and GDP that "change" information in the table. It is not fully functional that is why I would like to talk to you.

Still trying to change Polygons in MultyPolygons